refactor: express/vite serving code moved to server/
