+++ server/README.md
@@ -1,0 +1,4 @@
+NeoTavern-Frontend doesn't have any server logic.
+For that, see https://github.com/NeoTavern/NeoTavern-Server-Plugin
+
+But frontends still need to be served. This directory contains code that supports serving the frontend and its proxy to the SillyTavern backend.
