+++ server/vite-dev-overrides.ts
@@ -9,6 +9,7 @@ import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import type { ProxyOptions } from 'vite';
 import { z } from 'zod';
+import { isErrnoException } from './typeguards.ts';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const CONFIG_PATH = path.join(__dirname, 'launcher-config.json');
@@ -127,15 +128,3 @@ export const defaultDevOverrides: DevOverrides = {
   server: { port: 3000, host: false },
   preview: { port: 4173, host: true },
 };
-
-// wtf TypeScript. https://stackoverflow.com/a/70887388
-export function isErrnoException(error: unknown): error is NodeJS.ErrnoException {
-  if (!(error instanceof Error)) return false;
-  const err = error as unknown as { [key: string]: unknown };
-  return (
-    (typeof err.errno === 'number' || typeof err.errno === 'undefined') &&
-    (typeof err.code === 'string' || typeof err.code === 'undefined') &&
-    (typeof err.path === 'string' || typeof err.path === 'undefined') &&
-    (typeof err.syscall === 'string' || typeof err.syscall === 'undefined')
-  );
-}
